Create separate context provider for interactions
