+++ packages/lib/src/vis/shared/AxisSystemProvider.tsx
@@ -1,10 +1,8 @@
-import { assertDefined, isDefined } from '@h5web/shared';
 import { useThree } from '@react-three/fiber';
 import type { PropsWithChildren } from 'react';
-import { useCallback, useMemo, useState } from 'react';
+import { useCallback, useMemo } from 'react';
 import { Matrix4, Vector2, Vector3 } from 'three';
 
-import type { Interaction } from '../../interactions/models';
 import type { AxisConfig } from '../models';
 import { getCanvasScale, getSizeToFit } from '../utils';
 import { AxisSystemContext } from './AxisSystemContext';
@@ -53,46 +51,6 @@ function AxisSystemProvider(props: PropsWithChildren<Props>) {
     [camera, cameraToHtmlMatrix]
   );
 
-  const [interactionMap] = useState(new Map<string, Interaction>());
-
-  const registerInteraction = useCallback(
-    (id: string, value: Interaction) => {
-      interactionMap.set(id, value);
-    },
-    [interactionMap]
-  );
-
-  const unregisterInteraction = useCallback(
-    (id: string) => {
-      interactionMap.delete(id);
-    },
-    [interactionMap]
-  );
-
-  const shouldInteract = useCallback(
-    (id: string, event: MouseEvent) => {
-      const registeredKeys = [...interactionMap.values()]
-        .map((params: Interaction) => params.modifierKey)
-        .filter(isDefined);
-
-      const params = interactionMap.get(id);
-      assertDefined(params, `Interaction ${id} is not registered.`);
-
-      const { disabled, modifierKey } = params;
-      if (disabled) {
-        return false;
-      }
-
-      if (modifierKey !== undefined) {
-        return event.getModifierState(modifierKey);
-      }
-
-      // Check that there is no conflicting interaction
-      return registeredKeys.every((key) => !event.getModifierState(key));
-    },
-    [interactionMap]
-  );
-
   return (
     <AxisSystemContext.Provider
       value={{
@@ -105,9 +63,6 @@ function AxisSystemProvider(props: PropsWithChildren<Props>) {
         dataToWorld,
         worldToHtml,
         floatingToolbar,
-        registerInteraction,
-        unregisterInteraction,
-        shouldInteract,
       }}
     >
       {children}
